Camera reworked OBJ file support added kd-tree construction optimised Vector class replaced
-Support for 3D interactive camera added.
--Supports FOV and zoom

-Support for models with ".obj" extension added.
--Added several .obj modes
--Basic game world constructed in blender and added as .obj

-Lightning and shading on the GPU added

-Custom RVector class replaced with float3 CUDA type.

-Custom RColor class replaced with float4 CUDA tupe.
--Where Red = float4.x Green = float4.y Blue = float4.z and Special = float4.w

-CPU implementation optimised for new kd-tree data stucture

-Triangle normals are now being computed in ray-triangle intersection function.

+++ RayEngine/cutil_math.h
@@ -234,7 +234,7 @@ inline __host__ __device__ float2 normalize(float2 v)
 // floor
 inline __host__ __device__ float2 floor(const float2 v)
 {
-	return make_float2(floor(v.x), floor(v.y));
+	return make_float2(floorf(v.x), floorf(v.y));
 }
 
 // reflect
